backend update from Raze
* moving large allocations off the stack
* use proper printf formatters for size_t and ptrdiff_t.
* adding some missing 'noexcept'.

+++ src/common/filesystem/include/resourcefile.h
@@ -177,7 +177,7 @@ public:
 
 	int GetEntryNamespace(uint32_t entry)
 	{
-		return (entry < NumLumps) ? Entries[entry].Namespace : ns_hidden;
+		return (entry < NumLumps) ? Entries[entry].Namespace : (int)ns_hidden;
 	}
 
 	int GetEntryResourceID(uint32_t entry)
